backend/requirements.txt
mediapipe>=0.10.0
opencv-python>=4.7.0
numpy>=1.22.0
pillow>=9.0.0
tensorflow>=2.8.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart>=0.0.6
pydantic==2.4.2
skin-tone-classifier
pandas>=2.2.3
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0